+++ examples/fine-grained/src/Nested/C.hs
@@ -1,0 +1,6 @@
+module Nested.C (Factor (..), factor) where
+
+newtype Factor = Factor Int
+
+factor :: Factor
+factor = Factor 2